Finish adding add-remove functions to display elements in the DOM
