+++ __tests__/task_manager.spec.js
@@ -1,5 +1,9 @@
+/**
+ * @jest-environment jsdom
+ */
 import { addTask, removeFromStorage } from '../src/modules/task_manager.js';
-// import { displayTask } from '../src/index.js';
+import displayTask from '../src/modules/display.js';
+import removeAllChildElements from '../src/modules/remove_dom.js';
 
 describe('Add function tests', () => {
   test('Adding a message and one item to the list', () => {
@@ -19,7 +23,7 @@ describe('Add function tests', () => {
 });
 
 describe('Remove function test', () => {
-  test('Removing an item to the list', () => {
+  test('Removing an item from the list', () => {
     const tasksList = [
       { description: 'First message', index: 1, completed: false },
       { description: 'Second message', index: 2, completed: true },
@@ -28,7 +32,7 @@ describe('Remove function test', () => {
     const output = [
       { description: 'Second message', index: 1, completed: true },
     ];
-    expect(removeFromStorage(index, tasksList)).toEqual(output);
+    expect(removeFromStorage(tasksList, 'index', index)).toEqual(output);
   });
   test('Removing an item from a list of three elements', () => {
     const tasksList = [
@@ -41,24 +45,125 @@ describe('Remove function test', () => {
       { description: 'Second message', index: 1, completed: true },
       { description: 'Third message', index: 2, completed: false },
     ];
-    expect(removeFromStorage(index, tasksList)).toEqual(output);
+    expect(removeFromStorage(tasksList, 'index', index)).toEqual(output);
   });
 });
 
-// describe('Add list element to the DOM', () => {
-//   test('Add one new item to the list', () => {
-//     const uList = document.createElement('ul');
-//     uList.id = 'list';
-//     const newTask = {
-//       description: 'First message',
-//       index: 1,
-//       completed: false,
-//     };
-//     const itemToDisplay = displayTask(newTask);
-//     uList.appendChild(itemToDisplay);
+describe('Add list elements to the DOM', () => {
+  test('Add one new item to the list', () => {
+    document.body.innerHTML =
+    '<div>' +
+    '  <ul></ul>' +
+    '</div>';
+    const newTask = {description: 'First message', index: 1, completed: false};
 
-//     const items = document.querySelectorAll('#list li');
+    displayTask(newTask);
 
-//     expect(items).toHaveLength(1);
-//   });
-// });
+    const items = document.querySelectorAll('ul li');
+
+    expect(items).toHaveLength(1);
+  });
+  test('Add two new items to the list', () => {
+    document.body.innerHTML =
+    '<div>' +
+    '  <ul></ul>' +
+    '</div>';
+    const newTask = {description: 'First message', index: 1, completed: false};
+
+    displayTask(newTask);
+    displayTask(newTask);
+
+    const items = document.querySelectorAll('ul li');
+
+    expect(items).toHaveLength(2);
+  });
+  test('Add three new items to the list', () => {
+    document.body.innerHTML =
+    '<div>' +
+    '  <ul></ul>' +
+    '</div>';
+    const newTask = {description: 'First message', index: 1, completed: false};
+
+    displayTask(newTask);
+    displayTask(newTask);
+    displayTask(newTask);
+
+    const items = document.querySelectorAll('ul li');
+
+    expect(items).toHaveLength(3);
+  }); 
+});
+
+describe('Remove list elements from the DOM', () => {
+  test('Remove the only item from the list', () => {
+    document.body.innerHTML =
+    `<div>
+       <ul>
+         <li></li>
+       </ul>
+     </div>`;
+    
+    const uList = document.querySelector('ul');
+    const items = document.querySelectorAll('ul li');
+    removeAllChildElements(uList);
+    const trimmedList = items.length - 1;
+    const newTask = {description: 'First message', index: 1, completed: false};
+    
+    for (let i = 0; i < trimmedList; i+=1) {
+      displayTask(newTask);
+    }
+
+    const newItems = document.querySelectorAll('ul li');
+
+    expect(newItems).toHaveLength(0);
+  });
+  test('Remove one item from a list of two items', () => {
+    document.body.innerHTML =
+    `<div>
+       <ul>
+         <li></li>
+         <li></li>
+       </ul>
+     </div>`;
+    
+    const uList = document.querySelector('ul');
+    const items = document.querySelectorAll('ul li');
+    removeAllChildElements(uList);
+    const trimmedList = items.length - 1;
+    const newTask = {description: 'First message', index: 1, completed: false};
+    
+    for (let i = 0; i < trimmedList; i+=1) {
+      displayTask(newTask);
+    }
+
+    const newItems = document.querySelectorAll('ul li');
+
+    expect(newItems).toHaveLength(1);
+  }); 
+  test('Remove one item from a list of three items', () => {
+    document.body.innerHTML =
+    `<div>
+       <ul>
+         <li></li>
+         <li></li>
+         <li></li>
+       </ul>
+     </div>`;
+    
+    const uList = document.querySelector('ul');
+    const items = document.querySelectorAll('ul li');
+    removeAllChildElements(uList);
+    const trimmedList = items.length - 1;
+    const newTask = {description: 'First message', index: 1, completed: false};
+    
+    for (let i = 0; i < trimmedList; i+=1) {
+      displayTask(newTask);
+    }
+
+    const newItems = document.querySelectorAll('ul li');
+
+    expect(newItems).toHaveLength(2);
+  }); 
+});
+
+
